🔧 refactor: Update imports to utilize shared types
- Changed imports in `page.tsx` and `navbar.tsx` to source the `PR` type from `../types/objects` for better code organization.
- This refactor enhances maintainability by centralizing type definitions and reducing redundancy in the codebase.
